Move tests::vectors to tests::timing
There are no test vectors in those tests.
