+++ zebrad/src/components/sync/tests.rs
@@ -1,1 +1,1 @@
-mod vectors;
+mod timing;
